pyproject.toml
[tool]
[tool.poetry]
name = "TMan"
version = "0.0.0"
homepage = "https://github.com/Danmou/TMan"
description = "TODO"
authors = ["Daniel Mouritzen <dmrtzn@gmail.com>"]
readme = "README.md"
license =  "GPL-3.0-only"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
]
packages = [
    { include = "tman" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.10.0,<4.0"

# Runtime dependencies
PySimpleGUI = "^4.49.0"
appdirs = "^1.4.4"
attrs = "^21.2.0"
click = "^8.0.1"
loguru = "^0.5.3"
pydantic = "^1.8.2"

# Test dependencies
black = {version = "^21.5b2", optional = true}
flake8 = {version = "^4.0.1", optional = true}
flake8-docstrings = {version = "^1.6.0", optional = true}
hypothesis = {version = "^6.23.2", optional = true}
isort = {version = "^5.8.0", optional = true}
mypy = {version = "^0.910", optional = true}
pytest = {version = "^6.2.4", optional = true}
pytest-cov = {version = "^3.0.0", optional = true}
types-appdirs = {version = "^1.4.1", optional = true}
types-attrs = {version = "^19.1.0", optional = true}
types-click = {version = "^7.1.7", optional = true}
types-setuptools = {version = "^57.4.2", optional = true}

# Dev dependencies
bump2version = {version = "^1.0.1", optional = true}
pip = {version = "^20.3.1", optional = true}
pre-commit = {version = "^2.12.0", optional = true}
toml = {version = "^0.10.2", optional = true}
tox = {version = "^3.20.1", optional = true}
twine = {version = "^3.3.0", optional = true}
virtualenv = {version = "^20.2.2", optional = true}

# Doc dependencies
mkdocs = {version = "^1.1.2", optional = true}
mkdocs-autorefs = {version = "^0.3.0", optional = true}
mkdocs-include-markdown-plugin = {version = "^3.2.3", optional = true}
mkdocs-material = {version = "^7.3.3", optional = true}
mkdocs-material-extensions = {version = "^1.0.1", optional = true}
mkdocstrings = {version = "^0.16.2", optional = true}

[tool.poetry.extras]
test = [
    "black",
    "flake8",
    "flake8-docstrings",
    "hypothesis",
    "isort",
    "mypy",
    "pytest",
    "pytest-cov",
    "types-appdirs",
    "types-attrs",
    "types-click",
    "types-setuptools",
]

dev = [
    "bump2version",
    "pip",
    "pre-commit",
    "toml",
    "tox",
    "twine",
    "virtualenv",
]

doc = [
    "mkdocs",
    "mkdocs-autorefs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocs-material-extensions",
    "mkdocstrings",
]

[tool.poetry.scripts]
tman = 'tman.cli:main'
tman-gui = 'tman.gui:main'

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
line_length = 120
multi_line_output = 1
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
